Added opportunity functionality

+++ Conway.CRM.WebUI/Pages/Contacts/ContactList.razor.cs
@@ -1,6 +1,7 @@
 ﻿using Conway.CRM.Application.Interfaces;
 using Conway.CRM.Domain.Entities;
 using Microsoft.AspNetCore.Components;
+using Radzen;
 using Radzen.Blazor;
 
 namespace Conway.CRM.WebUI.Pages.Contacts
@@ -9,6 +10,7 @@ namespace Conway.CRM.WebUI.Pages.Contacts
     {
         [Inject] protected IContactRepository ContactRepository { get; set; }
         [Inject] protected NavigationManager NavigationManager { get; set; }
+        [Inject] protected DialogService DialogService { get; set; }
 
         protected RadzenDataGrid<Contact> grid;
         protected List<Contact> Contacts;
@@ -35,9 +37,15 @@ namespace Conway.CRM.WebUI.Pages.Contacts
 
         protected async Task DeleteContact(Guid contactId)
         {
-            await ContactRepository.DeleteContactAsync(contactId);
-            Contacts = (await ContactRepository.GetAllContactsAsync()).ToList();
-            await grid.Reload();
+            var result = await DialogService.Confirm(
+                "Are you sure you want to delete this Contact?", "Confirm Delete",
+                new ConfirmOptions() { OkButtonText = "Yes", CancelButtonText = "No" });
+            if (result.HasValue && result.Value)
+            {
+                await ContactRepository.DeleteContactAsync(contactId);
+                await LoadDataAsync();
+                await grid.Reload();
+            }
         }
     }
 }
